requirements.txt
# Workflow Launcher dependencies

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Gmail Integration (existing)
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0

# Client Hub - Database
supabase>=2.0.0

# Client Hub - Utilities
python-dateutil>=2.8.2
pydantic>=2.0.0
python-dotenv>=1.0.0
